fix new request page
